0_referencia2_tavily/requirements.txt
streamlit>=1.32.0
openai>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.0
pandas>=2.0.0
torchvision
anthropic
